+++ js_modules/dagit/packages/core/src/assets/__tests__/AssetPartitions.test.tsx
@@ -4,15 +4,14 @@ import userEvent, {specialChars} from '@testing-library/user-event';
 import React from 'react';
 import {MemoryRouter, Route} from 'react-router-dom';
 
-import {AssetKeyInput} from '../graphql/types';
-
-import {AssetPartitionListProps} from './AssetPartitionList';
-import {AssetPartitions} from './AssetPartitions';
-import {AssetViewParams} from './AssetView';
+import {AssetKeyInput} from '../../graphql/types';
+import {AssetPartitionListProps} from '../AssetPartitionList';
+import {AssetPartitions} from '../AssetPartitions';
+import {AssetViewParams} from '../AssetView';
 import {
   SingleDimensionStaticPartitionHealthQuery,
   SingleDimensionTimePartitionHealthQuery,
-} from './PartitionHealthSummary.mocks';
+} from '../__fixtures__/PartitionHealthSummary.mocks';
 
 // This file must be mocked because useVirtualizer tries to create a ResizeObserver,
 // and the component tree fails to mount. We still want to test whether certain partitions
